Moved hook-related types to Hook.swift
